qolfixup

+++ alpha_cc/training/trainer.py
@@ -335,7 +335,7 @@ class Trainer:
                 epoch_value_loss += value_loss.mean().cpu().item() / len(test_dataloader)
                 epoch_policy_loss += policy_loss.mean().cpu().item() / len(test_dataloader)
                 epoch_entropy_loss += entropy_loss.mean().cpu().item() / len(test_dataloader)
-                epoch_is_internal += is_internal.mean().cpu().item() / len(test_dataloader)
+                epoch_is_internal += is_internal.float().mean().cpu().item() / len(test_dataloader)
             if self._summary_writer is not None:
                 self._summary_writer.add_scalar("eval/policy-loss", epoch_policy_loss, global_step=self._eval_step)
                 self._summary_writer.add_scalar("eval/value-loss", epoch_value_loss, global_step=self._eval_step)
